[trello-27]: Moved form util to ds directroy in HW module
